requirements.txt
werkzeug
cryptography>=42.0.0
Flask==2.2.5
Jinja2==3.1.2
pycryptodome==3.20.0